requirements.txt
Django==3.0.5
djangorestframework==3.11.0
Markdown==3.2.1
coreapi==2.3.3